lttoolbox/compile/Compile.java: Added a FIXME comment for a reminder.
tagger/PatternList.java: PatternStore is just a typedef in the C++ version, replaced PatternStore with an equivalent of that underlying type. Java doesn't have multimaps in the standard library, so I made it a Map<Integer, ArrayList<Arraylist<Integer>>>.

transfer/MatchExe.java: Commented out extra line in MatchExe(MatchExe) copy constructor, dunno why if we are making a full copy of the object we would want to assign a brand new, clear finals object. Also switched copy function from public to private (like it should be and is in the C++ version), renamed to _copy as a result.

transfer/MatchState.java: Copy constructor somehow ended up private instead of public, fixed that. Renamed copy to _copy to better reflect that it is a private function.

tagger/PatternStore.java: Deleted. Because of the changes I just made in PatternList, this class is no longer needed and didn't really have anything implemented in it yet, just a skeleton.


git-svn-id: file:///home/apertium/svn-mirror/apertium-sf-net/trunk/lttoolbox-java@23189 2c8c2297-680c-4500-a755-649f80e260dd