tagger/MorphoStream.java, tagger/PatternList.java, tagger/Tagger.java: removed commented-out code
LTExpandTest.java, RegressionTest.java, EndianDoubleUtilTest.java: Adjusted the temp file paths such that they show up in a tmp subdirectory off the main project directory by prepending them with a "." (this directory has already been setup as being in the svn:ignore property of that folder).

This is for Windows compatibility. (Must still have a copy of "diff" on your path, however, which doesn't come with Windows by default.) Though the file can be written using a /tmp/somefile pathname, diff on Windows can't find it properly (at least not with my configuration).

git-svn-id: file:///home/apertium/svn-mirror/apertium-sf-net/trunk/lttoolbox-java@23235 2c8c2297-680c-4500-a755-649f80e260dd