IT FINALLY WORKS!!!!! At least on the single-word test, now to try it on longer inputs. XD
The issue was a difference in behavior between how the Java and C++ versions were checking for EOF. The C++ version would use feof() to check, but that would return false even if the next read would be eof, because that hadn't happened yet. Whereas the Java version doesn't have a way to check for EOF directly, so you have to actually read to find the EOF. XD More details in the comments I added to MorphoStream.java.

git-svn-id: file:///home/apertium/svn-mirror/apertium-sf-net/trunk/lttoolbox-java@23597 2c8c2297-680c-4500-a755-649f80e260dd